Fix infinite loop bug in update_min_max on Windows
On Windows make sure we update the value of the parameter “expected”
of atomic_compare_exchange_weak_explicit, to match the API’s
functionality with the std library’s implementation. As currently we
just macro atomic_compare_exchange_weak_explicit, to
InterlockedCompareExchange64  and if it failed to set a value, would
not update the expected value for us. Thus, we have implemented
atomic_compare_exchange and macroed
atomic_compare_exchange_weak_explicit to it. Before this changed, if
the value was of ->min_value or ->max_value was changed between the
observed and InterlockedCompareExchange64 we would enter into an
infinite loop as we never updated the observed value, as the value
of expected always remained the same thus,
InterlockedCompareExchange64 would always fail to set the
desired value.

Change-Id: Ic8b76bea1dc35d187376b3952ecb5ab258e874c9