Fix -Wunused-result warning due to posix_memalign()
Fix compiler warning -Wunused-result, due to the return value of
posix_memaling() not being used. The return value posix_memalign() is a
status code, for which 0 is success. Thus, if the value is not 0 we
have failed to allocate a memory to the pointer. So we should
return a NULL pointer to the caller.

Change-Id: I872edf473bfcb2f1d9dbdf7b50b43a4a7c8b78b3