MB-40967: Change hdr_iter to 64 bit
hdr_iter takes the total_count of the histogram (h) at construction.
The count used in the histogram is a 64 bit count. When we have more
than std::numeric_limits<int32_t>::max() counts in the histogram the
total_count field of the hdr_iter will underflow when we create the
iterator and various >0 checks will check in and prevent iteration of
the histogram.

Correct this by using a 64 bit type total_count field in hdr_iter.

Change-Id: I09d81dc75def8538852d7fc2307e98e7c821bdfe